Started implementing node fixing.
This implementation is based on v0.3-GSE-final:R/gse2.R.
Not tested yet!

Overhead is in
   model-bisse.R
      make.cache.bisse()
      ll.bisse()
   diversitree-branches.R
      xxsse.ll()
Actual fixing is in
   diversitree-branches.R
      all.branches()

Only two states are allowed so far.  For more, make node.fixing elements
into vectors rather than scalars.

I had a brief worry that node fixing might affect the E's as well as the
D's.  But I don't think so -- it just affects prob(this clade, given
this state).

Something else to worry about: if the "fixing" is not to a single state
(e.g. a node is 30% 0 and 70% 1), is c(1-a, a) okay even with logging
the likelihoods and using lq underflow compensation?

Checking my node-fixing ASRs against Rich's would definitely be a good
idea.