gp-bisse-eqs.c seems good
Appears to work for any number of states, so could be renamed without
"bisse".

The C routines take a list of parameters from R:
   nstates
   lambda 3D array
   mu 1D array
   q 2D array
The punctuated and gradual components of lambda are both used.

Testing notes are all in gpsdd-work1.R.  Answers agree with bisse special
case, but don't have a comparison for punctuated change or more states.