Functions to flatten and list-ify parameter values.
Added flatten.pars.gpbisse() and listify.pars.gpbisse().
Now argnames takes nstates as an argument.

These new functions are needed because find.mle(), mcmc(), and constrain() take
parameter vectors.  It would eventually be more efficient for the C code to
take a parameter vector, but as it is now, the confusing indexing logic is
confined to flatten and listify.