Resolved geosse unit test difference.
The discrepancy described in the previous message was because I changed
root.geosse() to use root.p when condition.surv = T (as Rich now does
for the other models), but that change never got passed to Rich.  So in
the code Rich sent back, root.geosse() was reverted.  Now it's back to
the new treatment.

Testing status:
   With the current code
      runit-classe.R : everything passes