Added option to condition on number of transitions (Mkn)
The new argument condition.trans sets the minimum number of transitions
required for a dataset to be considered.

The correction is an approximation: it ignores back-transitions and
applies the threshold to each state separately.

Seems to work fine for constrained likelihood functions.  But when
using find.mle(), need to pass keep.func=F.  (Need to fix things
in set.defaults().  Maybe also for make.bd.)